requirements.txt
pixeltable
openai
python-dotenv
azure.search
azure.search.documents==11.6.0b2
pydub
azure.core
ffmpeg-python
moviepy
librosa
numpy
mcp
opencv-python
pillow
soundfile
requests
tenacity
Authlib
cyclopts
fastmcp
